Add Swagger Config and UserMgmtController

+++ src/main/java/com/felixlaura/bindings/User.java
@@ -17,5 +17,5 @@ public class User {
 
     private LocalDate dob;
 
-    private Long ssn;
+    private Integer ssn;
 }
